+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## Next Release
+
+- Adds a spinner on project and deployment screens showing when a deployment is in progress. The screen will refresh every 5 seconds to update the status
+
 ## v2.0.0 (2024-06-10)
 
 - Upgrades Laravel from 10 to 11